requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.2
beautifulsoup4==4.13.3
boto3==1.37.2
botocore==1.37.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.1
daphne==4.1.2
Django==5.1.6
django-debug-toolbar==6.0.0
django-dotenv==1.4.2
django-storages==1.14.5
django-user-agents==0.4.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
joblib==1.4.2
lxml==5.3.1
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pip-review==1.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.3
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
toml==0.10.2
tomli==2.2.1
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
user-agents==2.2.0
wheel==0.45.1
zope.interface==7.2

# Development tools
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
ruff==0.8.6
pre-commit==4.0.1
